project/requirements.txt
attrs==22.1.0
bcrypt==4.0.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
clickclick==20.10.2
colorama==0.4.5
connexion==2.14.1
dnspython==2.2.1
Flask==2.2.2
Flask-PyMongo==2.3.0
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
MarkupSafe==2.1.1
packaging==21.3
PyJWT==2.4.0
pymongo==4.2.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
six==1.16.0
swagger-ui-bundle==0.0.9
urllib3==1.26.12
Werkzeug==2.2.2
